Allow specifying a CMAKE_TOOLCHAIN_FILE in the environment
For cross-compiling scenarios it's often necessary to pass
CMAKE_TOOLCHAIN_FILE to cmake. In a large project like Servo
there might be a lot of cmake-using dependencies that will all
need to pass it, so pulling this from an environment variable
will let people doing cross-compiles specify this just once.