Call "cmake" via "emcmake"/"emmake" for Emscripten
Emscripten toolchain has "emcmake" helper which sets the proper
CMAKE_TOOLCHAIN_FILE and other CMake variables that get CMake
to produce build files for cross-compiling via Emscripten.

There is also "emmake" helper that sets the environment variables
for cross-compilation. It should not be necessary for CMake when
used with Makefiles and Ninja, but Emscripten recommends it anyway.