Revert canonicalizing the build directory and make the path absolute instead (#267)
Rather than canonicalising the build directory this simply makes it
absolute. Unfortunately the current MSRV doesn't allow us to use
`std::path::absolute` so for Windows I've added a direct call to the
Windows API.

Since the motivation for #200 is Windows, on non-Windows platforms this
is now a no-op.

Fixes https://github.com/cloudflare/boring/issues/414